+++ src/index.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import Main from './Main.jsx';
+ReactDOM.createRoot(document.getElementById('root')).render(<Main />);
